# sample/docs/swagger.yaml
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  schema.Demo:
    properties:
      code:
        description: 编号
        type: string
      created_at:
        description: 创建时间
        type: string
      creator:
        description: 创建者
        type: string
      memo:
        description: 备注
        type: string
      name:
        description: 名称
        type: string
      record_id:
        description: 记录ID
        type: string
      status:
        description: 状态(1:启用 2:停用)
        type: integer
    required:
    - code
    - name
    - status
    type: object
  schema.DemoQueryParam:
    properties:
      code:
        description: 编号
        type: string
      like_code:
        description: 编号(模糊查询)
        type: string
      like_name:
        description: 名称(模糊查询)
        type: string
      status:
        description: 状态(1:启用 2:停用)
        type: integer
    type: object
info:
  contact: {}
  license: {}
paths:
  /api/v1/demos:
    post:
      description: 向你说描述
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 创建数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.Demo'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建单个demo
      tags:
      - Demo
  /api/v1/demos/{id}:
    get:
      description: 向你说描述
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询获取单个demo数据
      tags:
      - Demo
  /api/v1/demos/list:
    post:
      description: 向你说描述
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.DemoQueryParam'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询获取demo分页列表
      tags:
      - Demo
swagger: "2.0"